Creating target folder for storing the ouput of conversion to xml

+++ scripts/toxml.sh
@@ -93,8 +93,13 @@ rm "$XML_OUTPUT"/test.xml
 
 # Archiving the new declaration ids file, and also making it the new current declaration id file, to be used for the next scraping.
 now=$(date +'%Y-%m-%d')
-mkdir -p "$BASEDIR/archive/declarationids/"
-cp "$BASEDIR/declarationids-"$now "$BASEDIR/archive/declarationids/"
 
-mkdir -p  "$BASEDIR/currentdeclarationids"
-mv "$BASEDIR/declarationids-"$now "$BASEDIR/currentdeclarationids"
+# this file does not seem to be created so we'll cmpy/move it only if it exists
+if [ -f "$BASEDIR/declarationids-"$now ]; then
+	mkdir -p "$BASEDIR/archive/declarationids/"
+	cp "$BASEDIR/declarationids-"$now "$BASEDIR/archive/declarationids/"
+	mv "$BASEDIR/declarationids-"$now "$BASEDIR/currentdeclarationids"
+fi
+
+
+
